Create Recipe Option
The user is now able to create a recipe with as many ingredients as he wants. A dynamic form was created for this using templates and Javascript. After the post request, all inputs are sent back and processed inside python to create a dictionary of ingredients, and also to calculate the total cost of the recipe. NOTE: The package "pint" was installed to handle unit conversions during the total cost calculation step.